scripts hardening

+++ quickstart/docker/utils.sh
@@ -64,24 +64,27 @@ create_user() {
   local participant=$5
   echo "create_user $userId $userName $party $participant" >&2
 
-  # KV check whether user exist if not then create
-  curl_check "http://$participant:7575/v2/users" "$token" "application/json" \
-    --data-raw '{
-      "user" : {
-          "id" : "'$userId'",
-          "isDeactivated": false,
-          "primaryParty" : "'$party'",
-          "identityProviderId": "",
-          "metadata": {
-             "resourceVersion": "",
-              "annotations": {
-                  "username" : "'$userName'"
-              }
-          }
-      },
-        "rights": [
-        ]
-    }' | jq -r .user.id
+  code=$(curl_status_code "http://$participant:7575/v2/users/$userId" "$token" "application/json")
+  if  [ "$code" == "400" ]; then
+    curl_check "http://$participant:7575/v2/users" "$token" "application/json" \
+      --data-raw '{
+        "user" : {
+            "id" : "'$userId'",
+            "isDeactivated": false,
+            "primaryParty" : "'$party'",
+            "identityProviderId": "",
+            "metadata": {
+               "resourceVersion": "",
+                "annotations": {
+                    "username" : "'$userName'"
+                }
+            }
+        },
+          "rights": [
+          ]
+      }' | jq -r .user.id
+  fi
+
 }
 
 delete_user() {
@@ -90,8 +93,10 @@ delete_user() {
   local participant=$3
   echo "delete_user $userId $participant" >&2
 
-  # KV check whether user exist if so then delete
-  curl_check "http://$participant:7575/v2/users/$userId" "$token" "application/json" -X DELETE
+  code=$(curl_status_code "http://$participant:7575/v2/users/$userId" "$token" "application/json")
+  if  [ "$code" == "200" ]; then
+    curl_check "http://$participant:7575/v2/users/$userId" "$token" "application/json" -X DELETE
+  fi
 }
 
 joinByChar() {
@@ -107,6 +112,7 @@ function grant_rights() {
   local rights=$4
   local participant=$5
   echo "grant_rights $userId $partyId $rights $participant" >&2
+
   read -ra rightsAsArr <<< "$rights"
   local rightsArr=()
   for right in "${rightsAsArr[@]}"; do
@@ -174,53 +180,6 @@ upload_dars() {
   done
 }
 
-#allocate_party_and_create_user() {
-#  local token=$1
-#  local userId=$2
-#  local participant=$3
-#
-#  echo "create_user $userId $participant" >&2
-#
-#  party=$(get_user_party "$token" "$userId" "$participant")
-#  if [ -n "$party" ] && [ "$party" != "null" ]; then
-#    echo $party
-#    return
-#  fi
-#
-#  party=$(allocate_party "$token" "$userId" "$participant")
-#
-#  if [ -n "$party" ] && [ "$party" != "null" ]; then
-#    curl_check "http://$participant:7575/v2/users" "$token" "application/json" \
-#      --data-raw '{
-#        "user" : {
-#            "id" : "'$userId'",
-#            "primaryParty" : "'$party'",
-#            "isDeactivated": false,
-#            "identityProviderId": ""
-#        },
-#          "rights": [
-#              {
-#                  "kind": {
-#                      "CanActAs": {
-#                          "value": {
-#                              "party": "'$party'"
-#                          }
-#                      },
-#                      "CanReadAs": {
-#                          "value": {
-#                              "party": "'$party'"
-#                          }
-#                      }
-#                  }
-#              }
-#          ]
-#      }' | jq -r .user.primaryParty
-#  else
-#    echo "Failed to allocate party for user $userId" >&2
-#    exit 1
-#  fi
-#}
-
 get_user_party() {
   local token=$1
   local user=$2
@@ -228,53 +187,6 @@ get_user_party() {
   echo "get_user_party $user $participant" >&2
   curl_check "http://$participant:7575/v2/users/$user" "$token" "application/json" | jq -r .user.primaryParty
 }
-
-#allocate_party() {
-#  local token=$1
-#  local partyIdHint=$2
-#  local participant=$3
-#
-#  echo "allocate_party $partyIdHint $participant" >&2
-#
-#  namespace=$(get_participant_namespace "$token" "$participant")
-#
-#  party=$(curl_check "http://$participant:7575/v2/parties/party?parties=$partyIdHint::$namespace" "$token" "application/json" |
-#    jq -r '.partyDetails[0].party')
-#
-#  if [ -n "$party" ] && [ "$party" != "null" ]; then
-#    echo "party exists $party" >&2
-#    echo $party
-#    return
-#  fi
-#
-#  curl_check "http://$participant:7575/v2/parties" "$token" "application/json" \
-#    --data-raw '{
-#      "partyIdHint": "'$partyIdHint'",
-#      "displayName" : "'$partyIdHint'",
-#      "identityProviderId": ""
-#    }' | jq -r .partyDetails.party
-#}
-
-#get_participant_namespace() {
-#  local token=$1
-#  local participant=$2
-#  echo "get_participant_namespace $participant" >&2
-#  curl_check "http://$participant:7575/v2/parties/participant-id" "$token" "application/json" |
-#    jq -r .participantId | sed 's/^participant:://'
-#}
-
-#onboard_wallet_user() {
-#  local token=$1
-#  local user=$2
-#  local party=$3
-#  local validator=$4
-#  echo "onboard_wallet_user $user $party $validator" >&2
-#  curl_check "http://$validator:5003/api/validator/v0/admin/users" "$token" "application/json" \
-#    --data-raw '{
-#      "party_id": "'$party'",
-#      "name":"'$user'"
-#    }'
-#}
 
 get_dso_party_id() {
   local token=$1
@@ -309,4 +221,24 @@ curl_check() {
   fi
 
   echo "$responseBody"
+}
+
+curl_status_code() {
+  local url=$1
+  local token=$2
+  local contentType=${3:-application/json}
+  shift 3
+  local args=("$@")
+  echo "$url" >&2
+  if [ ${#args[@]} -ne 0 ]; then
+    echo "${args[@]}" >&2
+  fi
+
+  response=$(curl -s -S -w "\n%{http_code}" "$url" \
+      -H "Authorization: Bearer $token" \
+      -H "Content-Type: $contentType" \
+      "${args[@]}"
+      )
+
+  echo "$response" | tail -n1 | tr -d '\r'
 }
